fix: consolidate duplicate functions following architecture guidelines
- Consolidate filterValidFunctions in similarity/shared/
- Move formatDate to utils/date-utils.ts as global utility
- Consolidate determineFunctionType in analyzers/shared/
- Simplify getFunction/findFunction duplication with alias
