+++ src/utils/date-utils.ts
@@ -1,0 +1,32 @@
+/**
+ * Format a timestamp as a human-readable relative date
+ * 
+ * @param timestamp - Unix timestamp in milliseconds
+ * @returns Human-readable date string (e.g., "2h ago", "3d ago", "12/25/2023")
+ */
+export function formatDate(timestamp: number): string {
+  const date = new Date(timestamp);
+  const now = new Date();
+  const diffMs = now.getTime() - date.getTime();
+
+  // Less than 1 hour ago
+  if (diffMs < 60 * 60 * 1000) {
+    const minutes = Math.floor(diffMs / (60 * 1000));
+    return minutes <= 1 ? 'just now' : `${minutes}m ago`;
+  }
+
+  // Less than 24 hours ago
+  if (diffMs < 24 * 60 * 60 * 1000) {
+    const hours = Math.floor(diffMs / (60 * 60 * 1000));
+    return `${hours}h ago`;
+  }
+
+  // Less than 7 days ago
+  if (diffMs < 7 * 24 * 60 * 60 * 1000) {
+    const days = Math.floor(diffMs / (24 * 60 * 60 * 1000));
+    return `${days}d ago`;
+  }
+
+  // More than 7 days ago - show date
+  return date.toLocaleDateString();
+}
